innovateIN48-backend/requirements.txt
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
flask-swagger-ui==3.36.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycparser==2.20
PyJWT==1.7.1
six==1.16.0
Werkzeug==2.0.1
